v1.2 Collectible Implementation and polished Movement Controller
I was able to get a simple collectible script going and polish the character movement so that taking turns and accelerating felt much better than before. I tried to implement a wave pushdown for the player but it is a bit buggy right now so I will work on that tommorow.